Clean up app API file structure
